Fix hotspots behavior on iOS on the sample tour demo.
The hotspots on the sample tour demo were broken on iOS because of a bad
interaction between the touch event handlers used by the view controls and
the click event handler used by the hotspot.

Preventing touch events on hotspots from bubbling up into the controls
fixes this issue. Since we're already doing the same for scroll events,
merge the logic into a single helper function to be called on all hotspot
elements.