Remove tmp/version.js and the associated dev/build script logic.
The existence of tmp/version.js, which is generated by the development
scripts, makes it impossible to require Marzipano when installed through
npm. The version file is of little utility and is mainly a holdover
from Marzipano's closed-source days, so get rid of it.

It remains possible to identify a build version by the file preamble.