Use gl-matrix to calculate the rectilinear projection.
Currently, we roll our own code to calculate the projection matrix
for the rectilinear view when the projection center is nonzero. It
turns out that gl-matrix already includes code for this, so leverage
it instead.

In addition, remove projectionCenter from the exported API and move
the remaining logic into the VR demo code. The WebVR API is experimental
and we shouldn't make a commitment to it yet.