Rename Stage#updateSize() to setSize() and pass in an explicit size.
Passing in an explicit size enables the off-page rendering use case,
where the dimensions of the stage's underlying element can't be obtained
from the DOM. For regular clients, Viewer#updateSize() takes care of
getting the dimensions from the DOM and passing them into the stage on
the window resize handler.

Also clean up the ancilliary size update logic for the Css and Flash
stages. An explicit pixel size was being set on the DOM element, but this
is not required since the constructor sets its dimensions relative to the
parent element.